Moved Document (renamed Entry) class and Menu class to seperate files
